+++ example/lib/src/home/home_example.dart
@@ -39,6 +39,7 @@ class _MainExampleState extends State<MainExample> {
   ValueNotifier<bool> advPickerNotifierActivation = ValueNotifier(false);
   ValueNotifier<bool> trackingNotifier = ValueNotifier(false);
   ValueNotifier<bool> showFab = ValueNotifier(true);
+  ValueNotifier<GeoPoint?> lastGeoPoint = ValueNotifier(null);
   Timer? timer;
 
   @override
@@ -61,19 +62,36 @@ class _MainExampleState extends State<MainExample> {
     controller.listenerMapLongTapping.addListener(() async {
       if (controller.listenerMapLongTapping.value != null) {
         print(controller.listenerMapLongTapping.value);
-        await controller.addMarker(controller.listenerMapLongTapping.value!,
-            markerIcon: MarkerIcon(
-              icon: Icon(
-                Icons.store,
-                color: Colors.brown,
-                size: 48,
-              ),
-            ));
+        await controller.addMarker(
+          controller.listenerMapLongTapping.value!,
+          markerIcon: MarkerIcon(
+            icon: Icon(
+              Icons.store,
+              color: Colors.brown,
+              size: 48,
+            ),
+          ),
+          angle: pi / 3,
+        );
       }
     });
-    controller.listenerMapSingleTapping.addListener(() {
+    controller.listenerMapSingleTapping.addListener(() async{
       if (controller.listenerMapSingleTapping.value != null) {
+        if (lastGeoPoint.value != null) {
+          controller.removeMarker(lastGeoPoint.value!);
+        }
         print(controller.listenerMapSingleTapping.value);
+        lastGeoPoint.value = controller.listenerMapSingleTapping.value;
+        await controller.addMarker(
+          lastGeoPoint.value!,
+          markerIcon: MarkerIcon(
+            icon: Icon(
+              Icons.person,
+              color: Colors.red,
+            ),
+          ),
+          angle: -pi / 4,
+        );
       }
     });
 
@@ -85,9 +103,9 @@ class _MainExampleState extends State<MainExample> {
       // Future.delayed(Duration(seconds: 5), () async {
       //   await controller.zoomIn();
       // });
-      timer = Timer(Duration(seconds: 10), () async {
+      timer = Timer(Duration(seconds: 3), () async {
         await controller.setZoom(zoomLevel: 12);
-        await controller.setMarkerOfStaticPoint(
+        /*await controller.setMarkerOfStaticPoint(
           id: "line 2",
           markerIcon: MarkerIcon(
             icon: Icon(
@@ -111,6 +129,32 @@ class _MainExampleState extends State<MainExample> {
             ),
           ],
           "line 2",
+        );*/
+       await controller.addMarker(
+          GeoPoint(
+            latitude: 47.4517782,
+            longitude: 8.4716146,
+          ),
+          markerIcon: MarkerIcon(
+            icon: Icon(
+              Icons.person,
+              color: Colors.red,
+            ),
+          ),
+          angle: -pi / 4,
+        );
+        await controller.addMarker(
+          GeoPoint(
+            latitude: 47.4433594,
+            longitude: 8.4680184,
+          ),
+          markerIcon: MarkerIcon(
+            icon: Icon(
+              Icons.local_hospital,
+              color: Colors.red,
+            ),
+          ),
+          angle: pi / 4,
         );
         timer?.cancel();
       });
@@ -259,7 +303,7 @@ class _MainExampleState extends State<MainExample> {
                         GeoPoint(latitude: 47.4317782, longitude: 8.4716146),
                       ],
                     ),
-                    StaticPositionGeoPoint(
+                    /*StaticPositionGeoPoint(
                       "line 2",
                       MarkerIcon(
                         icon: Icon(
@@ -272,7 +316,7 @@ class _MainExampleState extends State<MainExample> {
                         GeoPoint(latitude: 47.4433594, longitude: 8.4680184),
                         GeoPoint(latitude: 47.4517782, longitude: 8.4716146),
                       ],
-                    )
+                    )*/
                   ],
                   road: Road(
                     startIcon: MarkerIcon(