+++ lib/src/controller/map_controller.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_osm_interface/flutter_osm_interface.dart';
 import 'package:flutter_osm_plugin/src/controller/osm/osm_controller.dart';
@@ -83,6 +85,7 @@ class MapController extends BaseMapController {
   ///change  Marker of specific static points
   /// we need to global key to recuperate widget from tree element
   /// [id] : (String) id  of the static group geopoint
+  ///
   /// [markerIcon] : (MarkerIcon) new marker that will set to the static group geopoint
   Future<void> setMarkerOfStaticPoint({
     required String id,
@@ -135,6 +138,11 @@ class MapController extends BaseMapController {
   Future<void> disabledTracking() async {
     await osmBaseController.disabledTracking();
   }
+
+  @Deprecated(
+    "this method will be removed in 0.25.0,use callback `listenerMapSingleTapping` or `listenerMapLongTapping` "
+    "to listener to click on the map, and use `addMarker` to create marker in that specific location",
+  )
 
   /// pick Position in map
   Future<GeoPoint> selectPosition({
@@ -258,11 +266,12 @@ class MapController extends BaseMapController {
   Future<void> addMarker(
     GeoPoint p, {
     MarkerIcon? markerIcon,
+    double? angle,
   }) async {
-    await osmBaseController.addMarker(
-      p,
-      markerIcon: markerIcon,
-    );
+    if (angle != null) {
+      assert(angle >= -pi && angle <= pi, "angle should be between -pi and pi");
+    }
+    await osmBaseController.addMarker(p, markerIcon: markerIcon, angle: angle);
   }
 }