+++ lib/src/controller/osm/osm_controller.dart
@@ -19,8 +19,8 @@ class MobileOSMController extends IBaseOSMController {
   Timer? _timer;
 
   late double stepZoom = 1;
-  late int minZoomLevel = 2;
-  late int maxZoomLevel = 18;
+  late double minZoomLevel = 2;
+  late double maxZoomLevel = 18;
 
   MobileOSMController();
 
@@ -148,8 +148,8 @@ class MobileOSMController extends IBaseOSMController {
       );
       await Future.delayed(Duration(milliseconds: 300), () async {
         if (_osmFlutterState.dynamicMarkerKey.currentContext != null) {
-          await changeDefaultIconMarker(_osmFlutterState.dynamicMarkerKey);
-          _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = null;
+          await changeIconAdvPickerMarker(_osmFlutterState.dynamicMarkerKey);
+          //_osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = null;
         }
       });
     }
@@ -267,8 +267,8 @@ class MobileOSMController extends IBaseOSMController {
 
   @override
   Future<void> configureZoomMap(
-    int minZoomLevel,
-    int maxZoomLevel,
+    double minZoomLevel,
+    double maxZoomLevel,
     double stepZoom,
     double initZoom,
   ) async {
@@ -405,20 +405,37 @@ class MobileOSMController extends IBaseOSMController {
   Future<void> addMarker(
     GeoPoint p, {
     MarkerIcon? markerIcon,
+    double? angle,
   }) async {
     if (markerIcon != null &&
         (markerIcon.icon != null || markerIcon.image != null)) {
       if (markerIcon.icon != null) {
         _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value =
-            markerIcon.icon;
+            angle == null || (angle == 0.0)
+                ? markerIcon.icon
+                : Transform.rotate(
+                    angle: angle,
+                    child: markerIcon.icon,
+                  );
       } else if (markerIcon.image != null) {
-        _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = Image(
-          image: markerIcon.image!,
-        );
+        _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value =
+            angle == null || (angle == 0.0)
+                ? Image(
+                    image: markerIcon.image!,
+                  )
+                : Transform.rotate(
+                    angle: angle,
+                    child: Image(
+                      image: markerIcon.image!,
+                    ),
+                  );
       }
-      Future.delayed(Duration(milliseconds: 300), () async {
-        await osmPlatform.addMarker(_idMap, p,
-            globalKeyIcon: _osmFlutterState.dynamicMarkerKey);
+      await Future.delayed(Duration(milliseconds: 300), () async {
+        await osmPlatform.addMarker(
+          _idMap,
+          p,
+          globalKeyIcon: _osmFlutterState.dynamicMarkerKey,
+        );
       });
     } else {
       await osmPlatform.addMarker(_idMap, p);