+++ flutter_osm_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.5 :
+* add attribute `angle` to method `addMarker`
+## 0.1.4 :
+* change type of minZoomLevel and maxZoomLevel from int to double
 ## 0.1.3 :
 ## 0.1.2 :
 ## 0.1.1 :